proddiscount/requirements.txt
future==0.16.0
grpcio==1.22.0
grpcio-tools==1.22.0
psycopg2==2.8.3
